pyproject.toml
[tool.poetry]
name = "poputka"
version = "0.1.0"
description = ""
authors = ["Max Bazarov <bazaroffm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.0"}
sqlalchemy = "^2.0.20"
celery = "^5.3.1"
flower = "^2.0.1"
alembic = "^1.11.3"
asyncpg = "^0.28.0"
pytest-asyncio = "^0.21.1"
bcrypt = "^4.0.1"
httpx = "^0.24.1"
python-jose = "^3.3.0"
pytest = "^7.4.0"
gunicorn = "^21.2.0"
aioredis = "^2.0.1"
sentry-sdk = {extras = ["fastapi"], version = "^1.30.0"}
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
pydantic-settings = "^2.0.3"
python-json-logger = "^2.0.7"
greenlet = "^3.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
